requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.7.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
Pillow==10.4.0